+++ packages/themes/Aprila/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.72.4](https://bitbucket.org/stacc-flow/bento/compare/v8.72.3...v8.72.4) (2019-03-08)
+
+
+### Bug Fixes
+
+* **aprila-theme:** remove margin top from submit button ([f1cfd91](https://bitbucket.org/stacc-flow/bento/commits/f1cfd91))
+
+
+
+
+
+## [8.72.3](https://bitbucket.org/stacc-flow/bento/compare/v8.72.2...v8.72.3) (2019-03-08)
+
+
+### Bug Fixes
+
+* **aprila-theme:** fix spacing + phoneInput ([3b24ef4](https://bitbucket.org/stacc-flow/bento/commits/3b24ef4))
+
+
+
+
+
+# [8.72.0](https://bitbucket.org/stacc-flow/bento/compare/v8.71.3...v8.72.0) (2019-03-08)
+
+
+### Bug Fixes
+
+* **aprila-theme:** fix help button theming ([ebaeefe](https://bitbucket.org/stacc-flow/bento/commits/ebaeefe))
+
+
+
+
+
 # [8.71.0](https://bitbucket.org/stacc-flow/bento/compare/v8.70.1...v8.71.0) (2019-03-05)
 
 **Note:** Version bump only for package @staccx/aprila-theme